mcp-server-m365/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
msal==1.25.0
pydantic==2.4.2
python-multipart==0.0.6
loguru==0.7.2
azure-identity==1.15.0
azure-graphrbac==0.61.1 